removed logs and comments and tested the final functionality of feature, allowing owners to add/delete descriptions
