+++ src/components/Warnings/Warnings.css
@@ -135,3 +135,8 @@
   margin-top: 0.5em;
   text-align: center;
 }
+
+.modal__warning__deletion {
+  font-weight: bold;
+  font-size: 1.3rem;
+}
